requirements.txt
PySocks==1.6.8
python-dotenv==0.10.1
python-telegram-bot==11.1.0
requests==2.21.0